QS qorld raking

+++ src/cbscraper/GenericWebScraper.py
@@ -32,6 +32,27 @@ n_requests = 0
 
 # FUNCTIONS
 
+# Print iterations progress
+def printProgressBar (iteration, total, prefix = '', suffix = '', decimals = 1, length = 100, fill = '█'):
+    """
+    Call in a loop to create terminal progress bar
+    @params:
+        iteration   - Required  : current iteration (Int)
+        total       - Required  : total iterations (Int)
+        prefix      - Optional  : prefix string (Str)
+        suffix      - Optional  : suffix string (Str)
+        decimals    - Optional  : positive number of decimals in percent complete (Int)
+        length      - Optional  : character length of bar (Int)
+        fill        - Optional  : bar fill character (Str)
+    """
+    percent = ("{0:." + str(decimals) + "f}").format(100 * (iteration / float(total)))
+    filledLength = int(length * iteration // total)
+    bar = fill * filledLength + '-' * (length - filledLength)
+    print('\r%s |%s| %s%% %s' % (prefix, bar, percent, suffix), end = '\r')
+    # Print New Line on Complete
+    if iteration == total:
+        print()
+
 def csv2stata(infile):
     stata_exe = "C:\Program Files (x86)\Stata13\StataMP-64.exe"
 
@@ -64,10 +85,13 @@ def readJSONFile(file):
 
 
 # filename DOES INCLUDE EXTENSION
-def saveJSON(data, filename):
+def saveJSON(data, filename, overwrite=True):
+    if not overwrite and os.path.isfile(filename):
+        return False
     data_str = jsonpickle.encode(data)
     with open(filename, 'w', encoding="utf-8") as fileh:
         fileh.write(data_str)
+    return True
 
 
 def myTextStrip(str):
